Move `get_locales()` from `main.py` to `languages.py`
